Move RID server into v1 package
