+++ pkg/rid/server/v1/doc.go
@@ -1,0 +1,3 @@
+// Package v1 provides types and functions required for implementing
+// RID server for the v1 API.
+package v1
