+++ pkg/rid/server/v1/server_test.go
@@ -1,4 +1,4 @@
-package server
+package v1
 
 import (
 	"context"
