+++ pkg/rid/server/v1/subscription_handler.go
@@ -1,4 +1,4 @@
-package server
+package v1
 
 import (
 	"context"
